requirements.txt
Django
django-wsgiserver

django-taggit
git+https://github.com/feuervogel/django-taggit-templatetags.git

#celery
#django-celery
#django-kombu

#python-ldap
#django-auth-ldap
